fs: Fix bug with `realpath` in `mesche_fs_resolve_path`
Apparently on Ubuntu (which is used in GitHub Actions CI images), `realpath`
gives different behavior than the one in `glibc` on my development machine.
Somehow, using the same char buffer in the input and output arguments to
`realpath` causes the buffer to be overwritten such that the result is
incorrect.

Using an intermediate char buffer and copying it back to the original buffer
seems to resolve the issue successfully, despite being a possible perf drain
down the road.